requirements-train.txt
mlflow>=2.8.0
python-dotenv>=1.0.0
huggingface_hub>=0.20.0
dagshub>=0.3.0
scikit-learn>=1.4.0
numpy<2.0
pandas>=1.5.0
joblib>=1.2.0
matplotlib>=3.6.0
seaborn>=0.12.0
xgboost>=1.7.0
